Make image urls absolute
When image paths don't start with a leading slash, it's harder to know
what the resulting url is going to be. If we are at a `/albums/123` url,
the path `images/foo.png` is broken, but if we're at the root url, `/`,
the image is there. I think the intention here is to never fetch images
relative to the current page, so I'm adding leading slashes throughout
the codebase.